Add payment method tracking:
